rename BoolTable -> BoolAlgTable
